Remove extra test and add to general config validation test file
